dts: oneplus-cheeseburger: Remove all MTP board ARM CoreSight nodes
These may not prevent booting, but are meant for debugging and tracing
software running on the SoC and simply do not make sense to be enabled
for embedded production devices such as the OnePlus 5.

After further testing it seems some of these also caused the boot
failures if I didn't add "clk_ignore_unused" to cmdline as well.

Some details: https://patchwork.kernel.org/patch/11172187/